requirements.txt
Flask
requests
beautifulsoup4
pandas
sqlalchemy
pymysql
mysql-connector-python
riotwatcher